Remove application-wrapper
