--- app/templates/components/application-wrapper.hbs
@@ -1,60 +1,0 @@
-{{#if fullPage}}
-	{{outlet}}
-{{else}}
-	{{site-head
-		drawerContent=activeDrawerContent
-		drawerVisible=drawerVisible
-		globalNavigation=this.attrs.globalNavigation
-		isFandomAppSmartBannerVisible=isFandomAppSmartBannerVisible
-		isSearchPage=isSearchPage
-		shadow=siteHeadShadow
-		setDrawerContent=(action 'setDrawerContent')
-		shouldShowFandomAppSmartBanner=shouldShowFandomAppSmartBanner
-		smartBannerVisible=smartBannerVisible
-		themeBar=themeBar
-		toggleDrawer=this.attrs.toggleDrawer
-		toggleSmartBannerVisibility=attrs.toggleSmartBanner
-	}}
-	{{#wikia-drawer
-		shouldBeVisible=drawerVisible
-	}}
-		{{#if activeDrawerContent}}
-			{{component
-				drawerContentComponent
-				onEnterHandler=(action 'closeDrawer')
-				suggestionClickAction=(action 'closeDrawer')
-				setDrawerContent=(action 'setDrawerContent')
-				closeDrawer=(action 'closeDrawer')
-				outsideSuggestionsClickAction=(action 'closeDrawer')
-				loadRandomArticle=this.attrs.loadRandomArticle
-				globalNavigation=this.attrs.globalNavigation
-				goToSearchResults=this.attrs.goToSearchResults
-				noScroll=true
-				focusInput=true
-			}}
-		{{/if}}
-	{{/wikia-drawer}}
-	{{#unless shouldShowFandomAppSmartBanner}}
-		{{smart-banner-android
-			isVisible=smartBannerVisible
-			toggleVisibility=attrs.toggleSmartBanner
-		}}
-	{{/unless}}
-	<div class="page-wrapper row" lang={{language.content}} dir={{language.contentDir}}>
-		{{outlet}}
-	</div>
-	{{lightbox-wrapper
-		type=lightboxType
-		model=lightboxModel
-		isVisible=lightboxVisible
-		lightboxCloseButtonDelay=lightboxCloseButtonDelay
-		closeLightbox=attrs.closeLightbox
-		setQueryParam=attrs.setQueryParam
-	}}
-	{{wikia-in-your-lang}}
-	{{#if globalFooter}}
-		{{global-footer/global-footer model=globalFooter}}
-	{{/if}}
-{{/if}}
-{{alert-notifications alerts=alertNotifications}}
-{{wds-spinner active=isLoading}}
